update meteor, styling on admin view
