+++ .meteor/packages
@@ -4,25 +4,25 @@
 # 'meteor add' and 'meteor remove' will edit this file for you,
 # but you can also edit it by hand.
 
-meteor-base@1.1.0             # Packages every Meteor app needs to have
+meteor-base@1.3.0             # Packages every Meteor app needs to have
 mobile-experience@1.0.5       # Packages for a great mobile UX
-mongo@1.2.3                  # The database Meteor supports right now
+mongo@1.4.2                  # The database Meteor supports right now
 blaze-html-templates@1.0.4    # Compile .html files into Meteor Blaze views
 reactive-var@1.0.11           # Reactive variable for tracker
 tracker@1.1.3                 # Meteor's client-side reactive programming library
 
-standard-minifier-css@1.3.5   # CSS minifier run for production mode
-standard-minifier-js@2.1.2    # JS minifier run for production mode
-es5-shim@4.6.15               # ECMAScript 5 compatibility for older browsers
-ecmascript@0.8.3              # Enable ECMAScript2015+ syntax in app code
-shell-server@0.2.4            # Server-side component of the `meteor shell` command
+standard-minifier-css@1.4.0   # CSS minifier run for production mode
+standard-minifier-js@2.3.1    # JS minifier run for production mode
+es5-shim@4.7.0               # ECMAScript 5 compatibility for older browsers
+ecmascript@0.10.0              # Enable ECMAScript2015+ syntax in app code
+shell-server@0.3.1            # Server-side component of the `meteor shell` command
 
 insecure@1.0.7                # Allow all DB writes from clients (for prototyping)
 react-meteor-data
 aldeed:simple-schema
 twbs:bootstrap
 session@1.1.7
-http@1.2.12
+http@1.4.0
 meteorhacks:aggregate
 momentjs:moment
 themeteorchef:bert
