+++ .meteor/release
@@ -1,1 +1,1 @@
-METEOR@1.5.4.1
+METEOR@1.6.1
