+++ .meteor/versions
@@ -1,88 +1,91 @@
-aldeed:simple-schema@1.5.3
-allow-deny@1.0.9
-autoupdate@1.3.12
-babel-compiler@6.20.0
-babel-runtime@1.0.1
-base64@1.0.10
+aldeed:simple-schema@1.5.4
+allow-deny@1.1.0
+autoupdate@1.4.0
+babel-compiler@7.0.5
+babel-runtime@1.2.2
+base64@1.0.11
 binary-heap@1.0.10
 blaze@2.3.2
 blaze-html-templates@1.1.2
 blaze-tools@1.0.10
-boilerplate-generator@1.2.0
-caching-compiler@1.1.9
+boilerplate-generator@1.4.0
+caching-compiler@1.1.11
 caching-html-compiler@1.1.2
-callback-hook@1.0.10
-check@1.2.5
-ddp@1.3.1
-ddp-client@2.1.3
-ddp-common@1.2.9
-ddp-server@2.0.2
+callback-hook@1.1.0
+check@1.3.1
+ddp@1.4.0
+ddp-client@2.3.2
+ddp-common@1.4.0
+ddp-server@2.1.2
 deps@1.0.12
-diff-sequence@1.0.7
-dynamic-import@0.1.3
-ecmascript@0.8.3
-ecmascript-runtime@0.4.1
-ecmascript-runtime-client@0.4.3
-ecmascript-runtime-server@0.4.1
-ejson@1.0.14
-es5-shim@4.6.15
+diff-sequence@1.1.0
+dynamic-import@0.3.0
+ecmascript@0.10.5
+ecmascript-runtime@0.5.0
+ecmascript-runtime-client@0.6.2
+ecmascript-runtime-server@0.5.0
+ejson@1.1.0
+es5-shim@4.7.3
 fortawesome:fontawesome@4.4.0_1
 fourseven:scss@4.5.4
 geojson-utils@1.0.10
 hot-code-push@1.0.4
 html-tools@1.0.11
 htmljs@1.0.11
-http@1.2.12
-id-map@1.0.9
+http@1.4.0
+id-map@1.1.0
 insecure@1.0.7
-jquery@1.11.10
+jquery@1.11.11
 launch-screen@1.1.1
 livedata@1.0.18
-logging@1.1.19
+logging@1.1.20
 mdg:validation-error@0.5.1
-meteor@1.7.2
-meteor-base@1.1.0
+meteor@1.8.4
+meteor-base@1.3.0
 meteorhacks:aggregate@1.3.0
 meteorhacks:collection-utils@1.2.0
-minifier-css@1.2.16
-minifier-js@2.1.4
-minimongo@1.3.3
+minifier-css@1.3.1
+minifier-js@2.3.3
+minimongo@1.4.3
 mobile-experience@1.0.5
 mobile-status-bar@1.0.14
-modules@0.10.0
-modules-runtime@0.8.0
-momentjs:moment@2.19.1
-mongo@1.2.3
-mongo-dev-server@1.0.1
-mongo-id@1.0.6
+modules@0.11.5
+modules-runtime@0.9.2
+momentjs:moment@2.21.0
+mongo@1.4.5
+mongo-dev-server@1.1.0
+mongo-id@1.0.7
 mongo-livedata@1.0.12
-npm-mongo@2.2.33
+npm-mongo@2.2.34
 observe-sequence@1.0.16
-ordered-dict@1.0.9
-promise@0.9.0
-random@1.0.10
-react-meteor-data@0.2.15
-reactive-dict@1.1.9
+ordered-dict@1.1.0
+promise@0.10.2
+random@1.1.0
+react-meteor-data@0.2.16
+reactive-dict@1.2.0
 reactive-var@1.0.11
-reload@1.1.11
-retry@1.0.9
-routepolicy@1.0.12
+reload@1.2.0
+retry@1.1.0
+routepolicy@1.0.13
+server-render@0.3.0
 session@1.1.7
-shell-server@0.2.4
+shell-server@0.3.1
+shim-common@0.1.0
+socket-stream-client@0.1.0
 spacebars@1.0.15
 spacebars-compiler@1.1.3
-standard-minifier-css@1.3.5
-standard-minifier-js@2.1.2
+standard-minifier-css@1.4.1
+standard-minifier-js@2.3.2
 templating@1.3.2
 templating-compiler@1.3.3
 templating-runtime@1.3.2
 templating-tools@1.1.2
 themeteorchef:bert@2.1.3
-tmeasday:check-npm-versions@0.3.1
+tmeasday:check-npm-versions@0.3.2
 tracker@1.1.3
 twbs:bootstrap@3.3.6
 ui@1.0.13
 underscore@1.0.10
-url@1.1.0
-webapp@1.3.19
+url@1.2.0
+webapp@1.5.0
 webapp-hashing@1.0.9
