+++ imports/ui/Login.jsx
@@ -1,16 +1,25 @@
 import React, { Component, PropTypes } from 'react';
 import ReactDOM from 'react-dom';
 import { Reviews } from '../api/classes.js';
-import Update from './Update.jsx';
-// Form component to allow user to add a review for selected course.
-// Takes in a course ID.
-// validation uses react controlled form elements as described at https://goshakkk.name/instant-form-fields-validation-react/
+import Admin from './Admin.jsx';
+import "./css/Login.css";
+
+/*
+  Admin Interface Login Component. Requests a password from the user to access the Admin
+  Interface.
+
+  Takes the given password and validates it with the password stored in the database.
+  If the user enters the correct password, the Admin Interface is displayed using
+  the Admin component. Otherwise, an error message is displayed.
+
+*/
+
 export default class Login extends Component {
   constructor(props) {
     super(props);
 
-    //store all currently selected form values in the state.
-    //this will be the default state.
+    // input elements are controlled components. Store the value of the
+    // user's input password and its validation state.
     this.state = {
       pass: "",
       validated: false,
@@ -20,29 +29,33 @@ export default class Login extends Component {
     this.defaultState = this.state;
   }
 
-  //save the pass, trigger re-render
+  // Save the user input text from the text box in the local state.
+  // Called whenever the input element changes to trigger re-render.
   handlePassChange(event) {
     this.setState({ pass: event.target.value });
   }
 
-  // handle a form submission. This will either add the review to the database
-  // or return an error telling the user to try agian.
+  // Handle a form submission to compare the user input to the password
+  // stored in the local database. If the validation passes, changed the
+  // local state 'validated' to true, otherwise, give an error message.
   handleSubmit(event) {
+    // pause auto-submit
     event.preventDefault();
 
     //ensure all fields are filled out
     var pass = this.state.pass.trim();
-    if (pass.length > 0 && pass !== null) {
+    if (pass.length > 0 && pass !== null && this.validateInputs(pass)) {
       // call the vailidate funtion
       Meteor.call('vailidateAdmin', pass, (error, result) => {
         if (!error && result==1) {
-          // Success, set in statealidate
+          // Success, set 'validate' in local state
           newState = {
             pass: "",
             validated: true
           };
           this.setState(newState);
         } else {
+          // otherwise, clear the input and send error message.
           this.setState(this.defaultState);
           this.setState({message: "Incorrect Password"});
         }
@@ -50,40 +63,37 @@ export default class Login extends Component {
     }
   }
 
-  //check if the state variables in the review are valid
-  validateInputs(median, attend, text) {
-    //ensure there are no illegal characters
+  // Check if the provided password is a valid string and not malicious to
+  // the database.
+  validateInputs(text) {
+    // ensure there are no illegal characters
     var regex = new RegExp(/^(?=.*[A-Z0-9])[\w:;.,?$%*#@[\]!--{}/\\()"'\/$ ]+$/i)
-    console.log(this.state.postClicks);
-    errs = {
-      median: median === null || median === undefined,
-      attend: attend === null || attend === undefined,
-      textEmpty: this.state.postClicks > 0 && (text === null || text === undefined || text.length === 0),
-      text: text != null && text != undefined && text.length > 0 && !regex.test(text),
-      allFalse: false
-    }
-    errs.allTrue = !(errs.median || errs.attend || errs.text || errs.textEmpty);
-    return errs;
+    return regex.test(text);
   }
 
   render() {
-    //check to see if all inputs are valid. If not, disable the post button and
-    //add a border around inputs that need to be changed.
+    // if password was valid, show admin interface, otherwise ask for the password.
     if (this.state.validated) {
       return (
-        <Update />
+        <Admin />
       );
     } else {
       return (
-        <form onSubmit={this.handleSubmit.bind(this)} >
+        <div className="pushDown">
           <div className="panel panel-default">
-            Enter Admin Password
+            <div className="panel-heading">
+              <h3 className="panel-title">Enter Admin Password</h3>
+            </div>
             <div className="panel-body">
-              <input ref="input" type="text" value={this.state.pass} onChange={(event) => this.handlePassChange(event)} />
-              <div>{this.state.message}</div>
+              <form onSubmit={this.handleSubmit.bind(this)} >
+                <div className="input-group fullInput">
+                  <input type="text" className="form-control" ref="input" value={this.state.pass} onChange={(event) => this.handlePassChange(event)} />
+                </div>
+                <div>{this.state.message}</div>
+              </form>
             </div>
           </div>
-        </form>
+        </div>
       );
     }
   }
