+++ imports/ui/css/Admin.css
@@ -1,0 +1,12 @@
+.whiteBg {
+  background-color: white !important;
+}
+
+.separate-buttons {
+  padding: 5px;
+}
+
+.width-90 {
+  width: 90% !important;
+  margin: 0 auto;
+}
