+++ imports/ui/css/Login.css
@@ -1,0 +1,8 @@
+.pushDown {
+  margin: 0 auto;
+  width: 80%;
+}
+
+.fullInput {
+  display: inline;
+}
