+++ imports/ui/css/Review.css
@@ -79,3 +79,7 @@
 .reviewNumbers{
   margin-bottom: 8px;
 }
+
+.space-review-buttons {
+  margin-left: 5px;
+}
