+++ package.json
@@ -5,6 +5,7 @@
     "start": "meteor run"
   },
   "dependencies": {
+    "@babel/runtime": "^7.0.0-beta.42",
     "babel-runtime": "^6.20.0",
     "bootstrap": "^3.3.7",
     "history": "^4.7.2",
